[FEAT] New Issue - contentContainer 구현
기존 newIssue.js에 있던 content 부분 contentContainer.js로 이동
